# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MeasureX API
  version: 1.0.0
  description: >
    One REST API serving the MeasureX mobile app, web dashboard and future
    integrations. Generated from PRD §13.


    Conventions:
      - JWT access token (15 min) + refresh token (12 h shift).
      - Errors: `{ code, message, details }`.
      - Cursor pagination on list endpoints.
      - `Idempotency-Key` header required on every create.
      - All times are UTC ISO 8601.
      - Base units are integers: dimensions in mm, weight in g. Billing figures
        are server-computed and final; the mobile app only shows a preview.
      - Role + branch checks are enforced server-side on every endpoint.
servers:
  - url: https://api.measurex.example/api/v1
    description: Staging (Saudi region)
  - url: http://localhost:3000/api/v1
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /shipments:
    get:
      tags: [Shipments]
      summary: List shipments (filtered, cursor-paginated)
      parameters:
        - { in: query, name: branchId, schema: { type: string, format: uuid } }
        - { in: query, name: status, schema: { $ref: '#/components/schemas/ShipmentStatus' } }
        - { in: query, name: dateFrom, schema: { type: string, format: date-time } }
        - { in: query, name: dateTo, schema: { type: string, format: date-time } }
        - { in: query, name: confidence, schema: { $ref: '#/components/schemas/Confidence' } }
        - { in: query, name: method, schema: { $ref: '#/components/schemas/MeasurementMethod' } }
        - { in: query, name: flag, schema: { $ref: '#/components/schemas/FlagType' } }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of shipments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Shipment' }
  /shipments/{awb}:
    get:
      tags: [Shipments]
      summary: Shipment detail with packages and current versions
      parameters:
        - $ref: '#/components/parameters/AwbPath'
      responses:
        '200':
          description: Shipment detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShipmentDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /shipments/{awb}/complete:
    post:
      tags: [Shipments]
      summary: Complete a shipment (in_progress → completed)
      parameters:
        - $ref: '#/components/parameters/AwbPath'
      responses:
        '200':
          description: Updated shipment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Shipment' }
        '409': { $ref: '#/components/responses/Conflict' }
  /shipments/{awb}/reopen:
    post:
      tags: [Shipments]
      summary: Reopen a completed shipment (reason required). Team Leader/Admin.
      parameters:
        - $ref: '#/components/parameters/AwbPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonBody' }
      responses:
        '200':
          description: Updated shipment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Shipment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ---------------------------------------------------------------- Packages
...
  /packages/{id}/corrections:
    post:
      tags: [Packages]
      summary: Correct a package (new version, old kept; reason required)
      description: >
        Team Leader/Admin, within the package's branch scope. Appends a new
        immutable measurement_version and repoints the current version (the old
        version is never edited), recomputes shipment totals, and writes an audit
        entry with before/after. A dimensions-only correction carries over the
        previous weight; a manual weight requires weightReason.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CorrectionCreate' }
      responses:
        '201':
          description: New version created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Package' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Package is not active (void/superseded cannot be corrected)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /packages/{id}/void:
    post:
      tags: [Packages]
      summary: Void a package (wrong AWB; reason required). Excluded from totals.
      description: >
        Team Leader/Admin, within the package's branch scope. Sets status=void,
        keeps the package number out of the active sequence, recomputes totals,
        and audits before/after.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonBody' }
      responses:
        '200':
          description: Voided package
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Package' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Package is not active (already void/superseded)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ---------------------------------------------------------------- Photos
...
  /flags:
    post:
      tags: [Review]
      summary: Create a flag (e.g. Labour "Flag for Team Leader" → worker_flag)
      description: >
        Any authenticated role. A shipmentId or packageId is required; the type
        defaults to worker_flag. The type is also added to the shipment's flags
        array so it surfaces on the dashboard. Full triage is M4.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FlagCreate' }
      responses:
        '201':
          description: Flag created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Flag' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    get:
      tags: [Review]
      summary: List flags for the review queue
      parameters:
        - { in: query, name: branchId, schema: { type: string, format: uuid } }
        - { in: query, name: type, schema: { $ref: '#/components/schemas/FlagType' } }
        - { in: query, name: status, schema: { $ref: '#/components/schemas/FlagStatus' } }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of flags
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Flag' }
  /flags/{id}/resolve:
    post:
      tags: [Review]
      summary: Approve or dismiss a flag (note optional). Team Leader/Admin.
      description: >
        Branch-scoped. `approve` closes the flag as resolved, `dismiss` as
        dismissed; both record the note and an audit entry. When no open flag of
        that type remains on the shipment, the type is removed from the
        shipment's flags summary.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [approve, dismiss]
                note: { type: string }
      responses:
        '200':
          description: Updated flag
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Flag' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Flag is not open (already resolved/dismissed)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /remeasurements:
    get:
      tags: [Review]
      summary: List remeasurement requests in scope. Team Leader/Admin.
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [open, done, cancelled] } }
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Page of remeasurement requests
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/RemeasureRequest' }
    post:
      tags: [Review]
      summary: Request a remeasurement (reason required). Team Leader/Admin.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RemeasureCreate' }
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RemeasureRequest' }
  /remeasurements/{id}/cancel:
    post:
      tags: [Review]
      summary: Cancel an open remeasurement request
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Cancelled request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RemeasureRequest' }

  # ---------------------------------------------------------------- Admin CRUD
...
  /ext/shipments/{awb}:
    get:
      tags: [Integration]
      summary: External shipment lookup (Phase C; API-key auth). Not active now.
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/AwbPath'
      responses:
        '200':
          description: Shipment detail for an integrated caller
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShipmentDetail' }
        '501':
          description: Not implemented until Phase C
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    AwbPath:
      in: path
      name: awb
      required: true
      schema: { type: string, pattern: '^AY\d{11}$' }
      example: AY20435032852
    IdPath:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
    Limit:
      in: query
      name: limit
      schema: { type: integer, default: 50, maximum: 200 }
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Role or branch not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Invalid state transition or conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    Page:
      type: object
      properties:
        items: { type: array, items: {} }
        nextCursor: { type: string, nullable: true }
    ReasonBody:
      type: object
      required: [reason]
      properties:
        reason: { type: string }

    # ---- Enums ----
...
    ShipmentStatus: { type: string, enum: [in_progress, completed, remeasure_required] }
    PackageStatus: { type: string, enum: [active, superseded, void] }
    Confidence: { type: string, enum: [high, medium, low] }
    MeasurementMethod: { type: string, enum: [marker, ar, manual] }
    WeightSource: { type: string, enum: [scale, manual, none] }
    PhotoKind: { type: string, enum: [raw, annotated] }
...
    FlagStatus: { type: string, enum: [open, resolved, dismissed] }
    FlagType:
      type: string
      enum:
        - low_confidence
        - medium_confidence
        - manual_entry
        - manual_weight
        - weight_discrepancy
        - check_dimensions
        - possible_duplicate
        - worker_flag
    RemeasureReason:
      type: string
      enum:
        - low_confidence
        - incorrect_dimensions
        - bad_photo
        - device_issue
        - customer_dispute
        - manual_verification
        - other
    RemeasureStatus: { type: string, enum: [open, done, cancelled] }
...
    Shipment:
      type: object
      properties:
        id: { type: string, format: uuid }
        awb: { type: string }
        branchId: { type: string, format: uuid, nullable: true }
        clientId: { type: string, format: uuid, nullable: true }
        status: { $ref: '#/components/schemas/ShipmentStatus' }
        flags: { type: array, items: { $ref: '#/components/schemas/FlagType' } }
        totals: { $ref: '#/components/schemas/ShipmentTotals' }
        completedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ShipmentTotals:
      type: object
      description: Sum of active packages; recomputed on every change.
      properties:
        pieces: { type: integer }
        cbm: { type: number, description: m³ to 4 decimals }
        actualG: { type: integer }
        volumetricG: { type: integer }
        chargeableG: { type: integer }
    ShipmentDetail:
      allOf:
        - $ref: '#/components/schemas/Shipment'
        - type: object
          properties:
            packages:
              type: array
              items: { $ref: '#/components/schemas/Package' }
    Package:
      type: object
      properties:
        id: { type: string, format: uuid, description: Client-generated UUID }
        shipmentId: { type: string, format: uuid }
        packageNumber: { type: integer, nullable: true, description: Final, server-assigned }
        provisionalNumber: { type: integer, nullable: true }
        status: { $ref: '#/components/schemas/PackageStatus' }
        stationId: { type: string, format: uuid, nullable: true }
        deviceId: { type: string, format: uuid, nullable: true }
        measuredBy: { type: string, format: uuid, nullable: true }
        confirmedAt: { type: string, format: date-time, nullable: true }
        syncReceivedAt: { type: string, format: date-time, nullable: true }
        currentVersion: { $ref: '#/components/schemas/MeasurementVersion' }
        photos: { type: array, items: { $ref: '#/components/schemas/Photo' } }
...
    CorrectionCreate:
      type: object
      required: [lengthMm, widthMm, heightMm, reason]
      description: >
        New dimensions (base units) plus a mandatory reason. Omitting the weight
        fields carries over the previous version's weight; a manual weight
        requires weightReason. The server recomputes all billing.
      properties:
        lengthMm: { type: integer }
        widthMm: { type: integer }
        heightMm: { type: integer }
        actualWeightG: { type: integer, nullable: true }
        weightSource: { $ref: '#/components/schemas/WeightSource' }
        scaleId: { type: string, format: uuid, nullable: true }
        weightReason: { type: string, nullable: true, description: Required when weightSource = manual }
        method: { $ref: '#/components/schemas/MeasurementMethod' }
        confidence: { $ref: '#/components/schemas/Confidence' }
        confidenceDetail: { type: object, additionalProperties: true, nullable: true }
        reason: { type: string }
    MeasurementVersion:
      type: object
      description: Immutable. Billing figures are server-computed and final.
      properties:
        id: { type: string, format: uuid }
        packageId: { type: string, format: uuid }
        versionNo: { type: integer }
        lengthMm: { type: integer }
        widthMm: { type: integer }
        heightMm: { type: integer }
        actualWeightG: { type: integer, nullable: true }
        weightSource: { $ref: '#/components/schemas/WeightSource' }
        scaleId: { type: string, format: uuid, nullable: true }
        method: { $ref: '#/components/schemas/MeasurementMethod' }
        confidence: { $ref: '#/components/schemas/Confidence' }
        confidenceDetail: { type: object, additionalProperties: true, nullable: true }
        divisorUsed: { type: integer }
        billingLCm: { type: integer }
        billingWCm: { type: integer }
        billingHCm: { type: integer }
        cbm: { type: number }
        volumetricG: { type: integer }
        chargeableG: { type: integer }
        createdBy: { type: string, format: uuid, nullable: true }
        reason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    Photo:
      type: object
      properties:
        id: { type: string, format: uuid }
        packageId: { type: string, format: uuid }
        versionId: { type: string, format: uuid, nullable: true }
        kind: { $ref: '#/components/schemas/PhotoKind' }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        bytes: { type: integer, nullable: true }
...
    Flag:
      type: object
      properties:
        id: { type: string, format: uuid }
        shipmentId: { type: string, format: uuid, nullable: true }
        packageId: { type: string, format: uuid, nullable: true }
        type: { $ref: '#/components/schemas/FlagType' }
        status: { $ref: '#/components/schemas/FlagStatus' }
        resolvedBy: { type: string, format: uuid, nullable: true }
        note: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    FlagCreate:
      type: object
      description: At least one of shipmentId / packageId is required.
      properties:
        shipmentId: { type: string, format: uuid }
        packageId: { type: string, format: uuid }
        type:
          allOf: [{ $ref: '#/components/schemas/FlagType' }]
          default: worker_flag
        note: { type: string }
    RemeasureCreate:
      type: object
      required: [shipmentId, packageIds, reason]
      properties:
        shipmentId: { type: string, format: uuid }
        packageIds: { type: array, items: { type: string, format: uuid } }
        reason: { $ref: '#/components/schemas/RemeasureReason' }
        note: { type: string, description: Required when reason = other }
    RemeasureRequest:
      type: object
      properties:
        id: { type: string, format: uuid }
        shipmentId: { type: string, format: uuid }
        packageIds: { type: array, items: { type: string, format: uuid } }
        reason: { $ref: '#/components/schemas/RemeasureReason' }
        note: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/RemeasureStatus' }
        requestedBy: { type: string, format: uuid }
        doneBy: { type: string, format: uuid, nullable: true }
        closedAt: { type: string, format: date-time, nullable: true }

    # ---- Audit ----
...